+++ src/actionProcessor/index.ts
@@ -8,7 +8,6 @@ export * from './httpClientActionProcessor';
 export * from './intellijActionProcessor';
 export * from './jsActionProcessor';
 export * from './httpClientActionProcessor';
-export * from './refMetaActionProcessor';
 export * from './responseAsVariableActionProcessor';
 export * from './testMethod';
 export * from './variableActionProcessor';