+++ src/actionProcessor/testMethod.ts
@@ -1,26 +1,35 @@
 import { chalkInstance, scriptConsole } from '../logger';
-import { parseError } from '../utils';
+import { HttpRegion, TestResult } from '../models';
+import { getRegionName, parseError } from '../utils';
 
 export const testSymbols = {
   ok: '✓',
   error: '✖',
 };
 
-
-export function test(message: string, testMethod: () => void): void {
-  const chalk = chalkInstance();
-  if (typeof testMethod === 'function') {
-    try {
-      testMethod();
-      scriptConsole.info(chalk`{green ${testSymbols.ok} ${message || 'Test passed'}}`);
-    } catch (err) {
-      process.exitCode = 20;
-      const description = parseError(err);
-      if (description) {
-        scriptConsole.error(chalk`{red ${testSymbols.error} ${message || 'Test failed'} (${description.file}:${description.line}:${description.offset})}`);
-      } else {
-        scriptConsole.error(chalk`{red ${testSymbols.error} ${message || err.message || 'Test failed'}}`);
+export function testFactory(httpRegion: HttpRegion) {
+  return function test(message: string, testMethod: () => void): void {
+    const chalk = chalkInstance();
+    const testResult: TestResult = {
+      message,
+      result: true
+    };
+    if (!httpRegion.testResults) {
+      httpRegion.testResults = [];
+      scriptConsole.info(chalk`{bold Tests for ${getRegionName(httpRegion)}}`);
+    }
+    httpRegion.testResults.push(testResult);
+    if (typeof testMethod === 'function') {
+      try {
+        testMethod();
+        scriptConsole.info(chalk`{green ${testSymbols.ok} ${message || 'Test passed'}}`);
+      } catch (err) {
+        process.exitCode = 20;
+        testResult.result = false;
+        testResult.error = parseError(err);
+        scriptConsole.error(chalk`{red ${testSymbols.error} ${message || 'Test failed'} (${testResult.error.displayMessage})}`);
       }
     }
-  }
+  };
 }
+