+++ src/models/httpRegion.ts
@@ -2,6 +2,7 @@ import { HttpRegionAction } from './httpRegionAction';
 import { HttpRequest } from './httpRequest';
 import { HttpResponse } from './httpResponse';
 import { HttpSymbol } from './httpSymbol';
+import { TestResult } from './testResult';
 
 
 export interface HttpRegion{
@@ -12,6 +13,7 @@ export interface HttpRegion{
   source?: string;
   metaData: Record<string, string>;
   symbol: HttpSymbol;
+  testResults?: Array<TestResult>;
 }